spark-streaming/requirements.txt
kafka-python
psycopg2-binary
pandas
watchdog